requirements.txt
tkinter
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
google-generativeai==0.3.1
streamlit==1.31.0
flask==3.0.2
python-dotenv==1.0.1
